# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ENUMERA API",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "name": "ENUMERA",
            "url": "https://enumera.tech",
            "email": "enumera@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/cours": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Liste des cours",
                "description": "Retourne la liste des cours",
                "operationId": "1e958664e8f761cc671e219f58e2efc2",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cours"
                ],
                "summary": "Création d'un cours",
                "description": "Crée une nouveau cours et retourne le cours créé",
                "operationId": "af2ee7a8fad944b7401cb222e6b59d1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "abreviation",
                                    "type",
                                    "lecon"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Intitulé du cours"
                                    },
                                    "abreviation": {
                                        "type": "string",
                                        "example": "Intitulé du cours"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "pdf"
                                    },
                                    "lecon": {
                                        "type": "string",
                                        "example": "Slug de la leçon"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum lores ipomd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cours créé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cours/{slug}": {
            "get": {
                "tags": [
                    "Cours"
                ],
                "summary": "Récupère un cours par son slug",
                "description": "Retourne un cours",
                "operationId": "4fd1ce64c61240009065701bba8a654f",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du cours à récupérer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cours"
                ],
                "summary": "Modification d'un cours",
                "description": "Modifie un cours et retourne le cours modifié",
                "operationId": "74db7982f5d60d22a23b20aa9c06ce5d",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du cours à modifié",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "abreviation",
                                    "type",
                                    "lecon"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "example": "Intitulé du cours"
                                    },
                                    "abreviation": {
                                        "type": "string",
                                        "example": "Intitulé du cours"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "pdf"
                                    },
                                    "lecon": {
                                        "type": "string",
                                        "example": "Slug de la leçon"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum lores ipomd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cours modifié avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cours non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies ne sont pas valides"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cours"
                ],
                "summary": "Suppression d'un cours par son slug",
                "description": "Retourne le cours supprimé",
                "operationId": "9661fd5439739ec646c54fffc3e76d00",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "slug du cours à supprimer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cours supprimé avec succès"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Slug validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cours non trouvé"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}